Dedupllicated the logging methods from base page and mobile base page
