+++ page_objects/Base_Page.py
@@ -10,6 +10,7 @@ import time,logging,os,inspect
 from utils.Base_Logging import Base_Logging
 from .driverfactory import DriverFactory
 from .selenium_objects import Selenium_Objects
+from .logging_objects import Logging_Objects
 from page_objects import PageFactory
 from utils.stop_test_exception_util import Stop_Test_Exception
 import conf.remote_credentials
@@ -37,7 +38,7 @@ class Borg:
 # Get the Base URL from the conf file
 base_url = conf.base_url_conf
 
-class Base_Page(Borg, Selenium_Objects):
+class Base_Page(Borg, Selenium_Objects, Logging_Objects):
     "Page class that all page models can inherit from"
 
     def __init__(self,base_url):
@@ -87,11 +88,6 @@ class Base_Page(Borg, Selenium_Objects):
     def turn_off_highlight(self):
         "Turn off the highlighting feature"
         self.highlight_flag = False
-
-    def get_failure_message_list(self):
-        "Return the failure message list"
-        return self.failure_message_list
-
 
     def switch_page(self,page_name):
         "Switch the underlying class to the required Page"
@@ -216,12 +212,6 @@ class Base_Page(Borg, Selenium_Objects):
 
         return self.screenshot_dir
 
-
-    def set_log_file(self):
-        'set the log file'
-        self.log_name = self.testname + '.log'
-        self.log_obj = Base_Logging(log_file_name=self.log_name,level=logging.DEBUG)
-
     def set_rp_logger(self,rp_pytest_service):
         "Set the reportportal logger"
         self.rp_logger = self.log_obj.setup_rp_logging(rp_pytest_service)
@@ -289,11 +279,9 @@ class Base_Page(Borg, Selenium_Objects):
             self.driver.get(url)
         self.wait(wait_time)
 
-
     def get_current_url(self):
         "Get the current URL"
         return self.driver.current_url
-
 
     def get_page_title(self):
         "Get the current page title"
@@ -302,7 +290,6 @@ class Base_Page(Borg, Selenium_Objects):
     def get_current_window_handle(self):
         "Return the latest window handle"
         return self.driver.current_window_handle
-
 
     def set_window_name(self,name):
         "Set the name of the current window name"
@@ -524,15 +511,6 @@ class Base_Page(Borg, Selenium_Objects):
         self.reset()
 
 
-    def write(self,msg,level='info'):
-        "Log the message"
-        msg = str(msg)
-        self.msg_list.append('%-8s:  '%level.upper() + msg)
-        if self.browserstack_flag is True:
-            if self.browserstack_msg not in msg:
-                self.msg_list.pop(-1) #Remove the redundant BrowserStack message
-        self.log_obj.write(msg,level)
-
 
     def report_to_testrail(self,case_id,test_run_id,result_flag,msg=''):
         "Update Test Rail"
@@ -585,36 +563,6 @@ class Base_Page(Borg, Selenium_Objects):
 
         return result_flag
 
-    def success(self,msg,level='info',pre_format='PASS: '):
-        "Write out a success message"
-        if level.lower() == 'critical':
-            level = 'info'
-        self.log_obj.write(pre_format + msg,level)
-        self.result_counter += 1
-        self.pass_counter += 1
-
-    def failure(self,msg,level='info',pre_format='FAIL: '):
-        "Write out a failure message"
-        self.log_obj.write(pre_format + msg,level)
-        self.result_counter += 1
-        self.failure_message_list.append(pre_format + msg)
-        if level.lower() == 'critical':
-            self.teardown()
-            raise Stop_Test_Exception("Stopping test because: "+ msg)
-
-    def log_result(self,flag,positive,negative,level='info'):
-        "Write out the result of the test"
-        if level.lower() == "inverse":
-            if flag is True:
-                self.failure(positive,level="error")
-            else:
-                self.success(negative,level="info")
-        else:
-            if flag is True:
-                self.success(positive,level=level)
-            else:
-                self.failure(negative,level=level)
-
     def read_browser_console_log(self):
         "Read Browser Console log"
         log = None
@@ -651,27 +599,6 @@ class Base_Page(Borg, Selenium_Objects):
             result_flag = False
 
         return result_flag
-
-    def write_test_summary(self):
-        "Print out a useful, human readable summary"
-        self.write('\n\n************************\n--------RESULT--------\nTotal number of checks=%d'%self.result_counter)
-        self.write('Total number of checks passed=%d\n----------------------\n************************\n\n'%self.pass_counter)
-        self.write('Total number of mini-checks=%d'%self.mini_check_counter)
-        self.write('Total number of mini-checks passed=%d'%self.mini_check_pass_counter)
-        failure_message_list = self.get_failure_message_list()
-        if len(failure_message_list) > 0:
-            self.write('\n--------FAILURE SUMMARY--------\n')
-            for msg in failure_message_list:
-                self.write(msg)
-        if len(self.exceptions) > 0:
-            self.exceptions = list(set(self.exceptions))
-            self.write('\n--------USEFUL EXCEPTION--------\n')
-            for (i,msg) in enumerate(self.exceptions,start=1):
-                self.write(str(i)+"- " + msg)
-        self.make_gif()
-        if self.gif_file_name is not None:
-            self.write("Screenshots & GIF created at %s"%self.screenshot_dir)
-            self.write('************************')
 
     def start(self):
         "Overwrite this method in your Page module if you want to visit a specific URL"
